% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{listExercises}
\alias{listExercises}
\title{List available excercises for this course}
\usage{
listExercises()
}
\value{
A data.frame with exercise titles and file names.  Start an
  exercise using \code{\link{start}}.
}
\description{
List available excercises for this course
}

