% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{start}
\alias{start}
\title{Start new exercises for this course}
\usage{
start(data = "data")
}
\arguments{
\item{data}{A character string.  The directory where data files should be
  copied.  Defaults to "data".}
}
\description{
Calling \code{\link{start}} will create a local copy of any new .r or .rmd
files for exercises for this course and update your data subfolder with
related data files.
}
\details{
Existing files will NOT be overwriten.  If you want a fresh copy of an
exercise simply rename your old file and call \code{\link{start}} again.

See \code{\link{submit}} to submit your solutions for review.
}

