% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/student.R
\name{submit}
\alias{submit}
\title{Submit your solution to an exercise}
\usage{
submit(exercise)
}
\arguments{
\item{exercise}{A character string.  The name of the file to submit.  See
  \code{\link{listExercises}} for a list of excercise file names for this
  course, or \code{\link{status}} to see which exercises you have started but
  not submitted.}
}
\description{
Submit your solution to an exercise
}

