% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/instructor.R
\name{update}
\alias{update}
\title{Update database of student submittions (for Instructors)}
\usage{
update(students = readLines(studentList))
}
\arguments{
\item{students}{A character vector of student usernames.  Defaults to all.}
}
\description{
Scans students' project directories to check for newly (re)submitted
assignments.
}

