% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/instructor.R
\name{updateData}
\alias{updateData}
\title{Update database of student submittions (for Instructors)}
\usage{
updateData()
}
\description{
Scans students' project directories to check for newly (re)submitted
assignments.
}

