% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/student.R
\name{view}
\alias{view}
\title{Open the course website in the Viewer pane}
\usage{
view(url = config$templateData$baseURL)
}
\arguments{
\item{url}{A string containing the URL to open.  Defaults to the course
  homepage.}
}
\description{
Open the course website in the Viewer pane
}

