% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{view}
\alias{view}
\title{Open the course website in the Viewer pane}
\usage{
view(url = "http://rna.wlu.edu/bio297")
}
\arguments{
\item{url}{A string containing the URL to open.  Defaults to the course
  homepage.}
}
\description{
Open the course website in the Viewer pane
}

